MULE-12157: Non descriptive error message when connected operation doesn't have a provider
